fix(msvc): silence first batch of /W4 /WX diagnostics
First round of Windows CI whack-a-mole. Four classes of warning
surfaced as errors under /W4 /WX:

- C4244 narrowing (live, fixed in source): IANA_ENCODING_TYPE
  is an unscoped enum (int-width) and was being assigned to
  a uint16_t field of MetadataEntry. Added explicit static_cast
  with a short comment explaining why. One live site; the other
  grep hit is inside a dead `#if 0` block.

- C4996 strncpy deprecation (suppressed, compat-only): the
  compat shim mimics upstream's 2008-era C API and uses strncpy
  by design. Defined _CRT_SECURE_NO_WARNINGS PRIVATE on
  igtl_compat. Our own modern code uses std:: helpers and
  stays strict.

- C4456 local-hides-local-declaration (suppressed globally on
  MSVC): codegen emits nested `for (std::size_t i = ...)` loops
  where the inner reuses the outer name. Stylistic, not buggy.
  Proper fix is in the codegen templates (tracked); suppress
  for now so we don't have to regenerate every message file
  before shipping Windows support.

- C4458 local-hides-class-member (same treatment): Header::pack
  emits `const std::size_t body_size = 58;` that shadows the
  Header::body_size member. Same root cause, same resolution.

POSIX build still 36/36 green. Windows CI (commit cd34d01) was
the signal source for this fix batch.

+++ core-cpp/compat/src/igtlMessageBase.cxx
@@ -126,7 +126,9 @@ int MessageBase::Pack() {
         for (const auto& kv : m_MetaData) {
             rt::MetadataEntry e;
             e.key = kv.first;
-            e.value_encoding = kv.second.first;
+            // IANA_ENCODING_TYPE is an unscoped enum (int-width);
+            // value_encoding is uint16_t. Explicit narrowing cast.
+            e.value_encoding = static_cast<std::uint16_t>(kv.second.first);
             e.value.assign(kv.second.second.begin(),
                            kv.second.second.end());
             entries.push_back(std::move(e));
